+++ app.py
@@ -3,5 +3,6 @@
 
 from src.yt_summarizer import app
 
+# For local development only
 if __name__ == "__main__":
     app.run(host='0.0.0.0', port=5001, debug=True)